+++ src/contracts/types/GovernanceTypes.sol
@@ -3,7 +3,7 @@ pragma solidity ^0.8.16;
 
 struct GovernanceData {
     address governance;
-    uint16 callConfig; // bitwise
+    uint32 callConfig; // bitwise
     uint64 lastUpdate;
 }
